+++ tidy3d/plugins/mode/mode_solver.py
@@ -448,36 +448,65 @@ class ModeSolver(Tidy3dBaseModel):
         ]
         return np.stack(eps_tensor, axis=0)
 
+    def _tensorial_material_profile_modal_plane_tranform(
+        self, mat_data: ArrayComplex4D
+    ) -> ArrayComplex4D:
+        """For tensorial material response function such as epsilon and mu, pick and tranform it to
+        modal plane with normal axis rotated to z.
+        """
+        # get rid of normal axis
+        mat_tensor = np.take(mat_data, indices=[0], axis=1 + self.normal_axis)
+        mat_tensor = np.squeeze(mat_tensor, axis=1 + self.normal_axis)
+
+        # convert to into 3-by-3 representation for easier axis swap
+        flat_shape = np.shape(mat_tensor)  # 9 components flat
+        tensor_shape = [3, 3] + list(flat_shape[1:])  # 3-by-3 matrix
+        mat_tensor = mat_tensor.reshape(tensor_shape)
+
+        # swap axes to plane coordinates (normal_axis goes to z)
+        if self.normal_axis == 0:
+            # swap x and y
+            mat_tensor[[0, 1], :, ...] = mat_tensor[[1, 0], :, ...]
+            mat_tensor[:, [0, 1], ...] = mat_tensor[:, [1, 0], ...]
+        if self.normal_axis <= 1:
+            # swap x (normal_axis==0) or y (normal_axis==1) and z
+            mat_tensor[[1, 2], :, ...] = mat_tensor[[2, 1], :, ...]
+            mat_tensor[:, [1, 2], ...] = mat_tensor[:, [2, 1], ...]
+
+        # back to "flat" representation
+        mat_tensor = mat_tensor.reshape(flat_shape)
+
+        # construct to feed to mode solver
+        return mat_tensor
+
+    def _diagonal_material_profile_modal_plane_tranform(
+        self, mat_data: ArrayComplex4D
+    ) -> ArrayComplex3D:
+        """For diagonal material response function such as epsilon and mu, pick and tranform it to
+        modal plane with normal axis rotated to z.
+        """
+        # get rid of normal axis
+        mat_tensor = np.take(mat_data, indices=[0], axis=1 + self.normal_axis)
+        mat_tensor = np.squeeze(mat_tensor, axis=1 + self.normal_axis)
+
+        # swap axes to plane coordinates (normal_axis goes to z)
+        if self.normal_axis == 0:
+            # swap x and y
+            mat_tensor[[0, 1], :, ...] = mat_tensor[[1, 0], :, ...]
+        if self.normal_axis <= 1:
+            # swap x (normal_axis==0) or y (normal_axis==1) and z
+            mat_tensor[[1, 2], :, ...] = mat_tensor[[2, 1], :, ...]
+
+        # construct to feed to mode solver
+        return mat_tensor
+
     def _solver_eps(self, freq: float) -> ArrayComplex4D:
         """Diagonal permittivity in the shape needed by solver, with normal axis rotated to z."""
 
         # Get diagonal epsilon components in the plane
         eps_tensor = self._get_epsilon(freq)
-
-        # get rid of normal axis
-        eps_tensor = np.take(eps_tensor, indices=[0], axis=1 + self.normal_axis)
-        eps_tensor = np.squeeze(eps_tensor, axis=1 + self.normal_axis)
-
-        # convert to into 3-by-3 representation for easier axis swap
-        flat_shape = np.shape(eps_tensor)  # 9 components flat
-        tensor_shape = [3, 3] + list(flat_shape[1:])  # 3-by-3 matrix
-        eps_tensor = eps_tensor.reshape(tensor_shape)
-
-        # swap axes to plane coordinates (normal_axis goes to z)
-        if self.normal_axis == 0:
-            # swap x and y
-            eps_tensor[[0, 1], :, ...] = eps_tensor[[1, 0], :, ...]
-            eps_tensor[:, [0, 1], ...] = eps_tensor[:, [1, 0], ...]
-        if self.normal_axis <= 1:
-            # swap x (normal_axis==0) or y (normal_axis==1) and z
-            eps_tensor[[1, 2], :, ...] = eps_tensor[[2, 1], :, ...]
-            eps_tensor[:, [1, 2], ...] = eps_tensor[:, [2, 1], ...]
-
-        # back to "flat" representation
-        eps_tensor = eps_tensor.reshape(flat_shape)
-
-        # construct eps to feed to mode solver
-        return eps_tensor
+        # tranformation
+        return self._tensorial_material_profile_modal_plane_tranform(eps_tensor)
 
     def _solve_all_freqs(
         self,
@@ -498,6 +527,22 @@ class ModeSolver(Tidy3dBaseModel):
             eps_spec.append(eps_spec_freq)
 
         return n_complex, fields, eps_spec
+
+    def _postprocess_solver_fields(self, solver_fields):
+        """Postprocess `solver_fields` from `compute_modes` to proper coordinate"""
+        fields = {key: [] for key in ("Ex", "Ey", "Ez", "Hx", "Hy", "Hz")}
+        for mode_index in range(self.mode_spec.num_modes):
+            # Get E and H fields at the current mode_index
+            ((Ex, Ey, Ez), (Hx, Hy, Hz)) = self._process_fields(solver_fields, mode_index)
+
+            # Note: back in original coordinates
+            fields_mode = {"Ex": Ex, "Ey": Ey, "Ez": Ez, "Hx": Hx, "Hy": Hy, "Hz": Hz}
+            for field_name, field in fields_mode.items():
+                fields[field_name].append(field)
+
+        for field_name, field in fields.items():
+            fields[field_name] = np.stack(field, axis=-1)
+        return fields
 
     def _solve_single_freq(
         self,
@@ -522,19 +567,7 @@ class ModeSolver(Tidy3dBaseModel):
             direction=self.direction,
         )
 
-        fields = {key: [] for key in ("Ex", "Ey", "Ez", "Hx", "Hy", "Hz")}
-        for mode_index in range(self.mode_spec.num_modes):
-            # Get E and H fields at the current mode_index
-            ((Ex, Ey, Ez), (Hx, Hy, Hz)) = self._process_fields(solver_fields, mode_index)
-
-            # Note: back in original coordinates
-            fields_mode = {"Ex": Ex, "Ey": Ey, "Ez": Ez, "Hx": Hx, "Hy": Hy, "Hz": Hz}
-            for field_name, field in fields_mode.items():
-                fields[field_name].append(field)
-
-        for field_name, field in fields.items():
-            fields[field_name] = np.stack(field, axis=-1)
-
+        fields = self._postprocess_solver_fields(solver_fields)
         return n_complex, fields, eps_spec
 
     def _rotate_field_coords(self, field: FIELD) -> FIELD: