+++ tidy3d/plugins/smatrix/__init__.py
@@ -1,5 +1,17 @@
 """ Imports from scattering matrix plugin. """
 
-from .smatrix import ComponentModeler, Port, SMatrixDataArray
+from .ports.modal import Port
+from .ports.lumped import LumpedPort
+from .component_modelers.modal import AbstractComponentModeler
+from .component_modelers.modal import ComponentModeler, ModalPortDataArray
+from .component_modelers.terminal import TerminalComponentModeler, LumpedPortDataArray
 
-__all__ = ["ComponentModeler", "Port", "SMatrixDataArray"]
+__all__ = [
+    "AbstractComponentModeler",
+    "ComponentModeler",
+    "Port",
+    "ModalPortDataArray",
+    "TerminalComponentModeler",
+    "LumpedPort",
+    "LumpedPortDataArray",
+]